backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
boto3==1.34.0
librosa==0.10.1
madmom==0.16.1
aubio==0.4.9
soundfile==0.12.1
pydub==0.25.1
openl3==0.4.1
requests==2.31.0
google-generativeai==0.3.2
pronouncing==0.2.0
pyphen==0.14.0
nltk==3.8.1
reportlab==4.0.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
yt-dlp==2023.12.30
